Move DefaultErrorResponse.kt into model.error package.
