+++ app/src/main/java/com/appunite/loudius/network/model/error/DefaultErrorResponse.kt
@@ -1,4 +1,4 @@
-package com.appunite.loudius.network.utils
+package com.appunite.loudius.network.model.error
 
 data class DefaultErrorResponse(
     val message: String,
