+++ app/src/main/java/com/appunite/loudius/network/utils/ApiCallUtil.kt
@@ -1,9 +1,10 @@
 package com.appunite.loudius.network.utils
 
+import com.appunite.loudius.network.model.error.DefaultErrorResponse
 import com.google.gson.Gson
+import java.io.IOException
 import org.json.JSONException
 import retrofit2.HttpException
-import java.io.IOException
 
 suspend fun <T> safeApiCall(
     errorParser: RequestErrorParser = DefaultErrorParser,
